Update Sphinx navigation setting and modify IMAP documentation
This commit primarily adjusts the Sphinx navigation depth in the configuration file and revises the Python documentation relating to IMAP. Various parts of the IMAP documentation were modified, from amending headers and context to explaining the imaplib.IMAP4_SSL function in greater detail. Simultaneously, the Jupiter section of the Python documentation and all the related files were removed entirely.