Update and include more details in IMAP documentation
This commit updates the IMAP documentation by adding detailed descriptions of the `fetch` and `search` functions in the IMAP protocol, which are used for retrieving specific mail data and searching for messages in a mailbox, respectively. Syntax modifications were made for method representations to align them with Sphinx's Python Domain format.